requirements.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
./dlib-19.22.99-cp38-cp38-win_amd64.whl
django==3.2.9
Cython==0.29.28
easydict==1.10
grpcio
matplotlib==3.5.3
matplotlib-inline==0.1.6
numpy~=1.24.4
oauthlib==3.2.2
opencv-python==4.6.0.66
pandas==1.3.5
peewee==3.15.4
Pillow~=10.4.0
protobuf==3.20.3
psutil==5.9.4
ipython
PyYAML==6.0
seaborn
requests==2.28.1
requests-oauthlib==1.3.1
scipy~=1.10.1
six
smart-open==6.2.0
sqlparse==0.4.3
tqdm==4.64.1
onnx==1.13.0
nncf==2.7.0
openvino==2023.1

keras~=2.13.1
tensorflow~=2.13.0
scikit-learn~=1.3.2
dlib~=19.22.99
diGraph~=0.1.4